Depth Quality Tool fixes: Make Z-Accuracy sign-aware fill-rate shall apear regardless of plane-fit status

+++ common/ux-window.cpp
@@ -78,7 +78,7 @@ namespace rs2
 
         // Prepare the splash screen and do some initialization in the background
         int x, y, comp;
-        auto r = stbi_load_from_memory(splash, splash_size, &x, &y, &comp, false);
+        auto r = stbi_load_from_memory(splash, (int)splash_size, &x, &y, &comp, false);
         _splash_tex.upload_image(x, y, r);
 
         // Apply initial UI state
@@ -131,7 +131,7 @@ namespace rs2
             begin_frame();
 
             glPushMatrix();
-            glViewport(0.f, 0.f, _fb_width, _fb_height);
+            glViewport(0, 0, _fb_width, _fb_height);
             glClearColor(0.036f, 0.044f, 0.051f, 1.f);
             glClear(GL_COLOR_BUFFER_BIT);
